Fix implementation so that it compiles correctly
